order entries correctly
If possible convert last char of book identifier to integer (assuming that e.g. book 3, vol 3 or 101.3 would valid forms for identifying books). If it is an integer multiply by 10k and add page number. But convert result to string for comparability with those cases where book number final character or page number is not easily translated to integer